Date-related refactors, also included a factory expect/actual method
Well, no point in dates being in the koinModules package; and platform-specific
classes have the 'actual' createDate method implementations too
