+++ composeApp/src/commonMain/kotlin/date/BaseDate.kt
@@ -1,0 +1,7 @@
+package date
+
+interface BaseDate {
+    fun format(pattern: String): String
+}
+
+expect fun createDate(year: Int, month: Int, day: Int): BaseDate
