--- composeApp/src/commonMain/kotlin/koinModules/interface/BaseDate.kt
@@ -1,5 +1,0 @@
-package koinModules.`interface`
-
-interface BaseDate {
-    fun format(pattern: String): String
-}
