+++ composeApp/src/iosMain/kotlin/date/IOSDate.kt
@@ -1,9 +1,11 @@
-package koinModules
-
-import koinModules.`interface`.BaseDate
+package date
 
 class IOSDate: BaseDate {
     override fun format(pattern: String): String {
         TODO("Not yet implemented")
     }
+}
+
+actual fun createDate(year: Int, month: Int, day: Int): BaseDate {
+    TODO("Not implemented")
 }
